requirements.txt
discord.py
python-dotenv
discord-pretty-help
aiosqlite
dblpy